+++ R/spp_cites_legislation.R
@@ -24,7 +24,7 @@
 #' @param ... Further named parameters, see [httr::GET()].
 #'
 #' @return If `raw` is set to `TRUE` then an object of class `spp_raw` (or
-#' `spp_raw_multi` if `length(taxon_id)>1`) is returned which is essentially
+#' `spp_raw_multi` if `length(taxon_id) > 1`) is returned which is essentially
 #' a list of lists (see option `as = 'parsed'` in [httr::content()]).
 #' Otherwise, an object of class `spp_cites_leg` (or `spp_cites_leg_multi` if
 #' `length(taxon_id)>1`) is returned which is a list of three data frames: